[Security] Revert changes made between 2.7 and 2.8-beta
