+++ src/Symfony/Component/Security/Core/Tests/Authorization/Voter/AbstractVoterTest.php
@@ -11,10 +11,11 @@
 
 namespace Symfony\Component\Security\Core\Tests\Authorization\Voter;
 
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\Authorization\Voter\AbstractVoter;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 
+/**
+ * @group legacy
+ */
 class AbstractVoterTest extends \PHPUnit_Framework_TestCase
 {
     protected $token;
@@ -50,75 +51,8 @@ class AbstractVoterTest extends \PHPUnit_Framework_TestCase
      */
     public function testVote(array $attributes, $expectedVote, $object, $message)
     {
-        $voter = new AbstractVoterTest_Voter();
+        $voter = new Fixtures\MyVoter();
 
         $this->assertEquals($expectedVote, $voter->vote($this->token, $object, $attributes), $message);
     }
-
-    /**
-     * @dataProvider getTests
-     * @group legacy
-     */
-    public function testVoteLegacy(array $attributes, $expectedVote, $object, $message)
-    {
-        $voter = new AbstractVoterTest_LegacyVoter();
-
-        $this->assertEquals($expectedVote, $voter->vote($this->token, $object, $attributes), $message);
-    }
-
-    /**
-     * @group legacy
-     * @expectedException \BadMethodCallException
-     */
-    public function testNoOverriddenMethodsThrowsException()
-    {
-        $voter = new AbstractVoterTest_NothingImplementedVoter();
-        $voter->vote($this->token, new \stdClass(), array('EDIT'));
-    }
-}
-
-class AbstractVoterTest_Voter extends AbstractVoter
-{
-    protected function voteOnAttribute($attribute, $object, TokenInterface $token)
-    {
-        return 'EDIT' === $attribute;
-    }
-
-    protected function supports($attribute, $object)
-    {
-        return $object instanceof \stdClass && in_array($attribute, array('EDIT', 'CREATE'));
-    }
-}
-
-class AbstractVoterTest_LegacyVoter extends AbstractVoter
-{
-    protected function getSupportedClasses()
-    {
-        return array('stdClass');
-    }
-
-    protected function getSupportedAttributes()
-    {
-        return array('EDIT', 'CREATE');
-    }
-
-    protected function isGranted($attribute, $object, $user = null)
-    {
-        return 'EDIT' === $attribute;
-    }
-}
-
-class AbstractVoterTest_NothingImplementedVoter extends AbstractVoter
-{
-    protected function getSupportedClasses()
-    {
-        return array('stdClass');
-    }
-
-    protected function getSupportedAttributes()
-    {
-        return array('EDIT', 'CREATE');
-    }
-
-    // this is a bad voter that hasn't overridden isGranted or voteOnAttribute
 }
